+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-07-11 16:57:07
-// Commit: style(i18n): harmonize light and dark theme palette for i18n
+// Updated at: 2026-07-11 21:42:48
+// Commit: test(i18n): add unit tests for i18n business logic
 
 class i18nComponentState {
-  final String commitRef = '2700';
-  final String lastMessage = 'style(i18n): harmonize light and dark theme palette for i18n';
-  final DateTime timestamp = DateTime.parse('2026-07-11T16:57:07+05:00');
+  final String commitRef = '2716';
+  final String lastMessage = 'test(i18n): add unit tests for i18n business logic';
+  final DateTime timestamp = DateTime.parse('2026-07-11T21:42:48+05:00');
 }